+++ src/Control/Monad/Ether.hs
@@ -2,22 +2,13 @@
 -- | This module provides convenience exports of all
 -- tagged monad classes from Ether.
 
-module Control.Monad.Ether
-    ( module Control.Monad
-    , module Control.Monad.Fix
-    , module Control.Monad.Ether.Reader
-    , module Control.Monad.Ether.Writer
-    , module Control.Monad.Ether.State
-    , module Control.Monad.Ether.Except
-    , module Control.Ether.Wrapped
-    , ethereal
-    ) where
+module Control.Monad.Ether (module X) where
 
-import Control.Monad
-import Control.Monad.Fix
-import Control.Monad.Ether.Reader
-import Control.Monad.Ether.Writer
-import Control.Monad.Ether.State
-import Control.Monad.Ether.Except
-import Control.Ether.Wrapped
-import Control.Ether.TH (ethereal)
+import Control.Monad as X
+import Control.Monad.Fix as X
+import Control.Monad.Ether.Reader as X
+import Control.Monad.Ether.Writer as X
+import Control.Monad.Ether.State as X
+import Control.Monad.Ether.Except as X
+import Control.Monad.Trans.Ether.Dispatch as X hiding (pack, unpack)
+import Control.Ether.TH as X (ethereal)