+++ src/Control/Monad/Ether/Implicit.hs
@@ -2,18 +2,13 @@
 -- | This module provides convenience exports of all
 -- implicitly tagged monad classes from Ether.
 
-module Control.Monad.Ether.Implicit
-    ( module Control.Monad
-    , module Control.Monad.Fix
-    , module Control.Monad.Ether.Implicit.Reader
-    , module Control.Monad.Ether.Implicit.Writer
-    , module Control.Monad.Ether.Implicit.State
-    , module Control.Monad.Ether.Implicit.Except
-    ) where
+module Control.Monad.Ether.Implicit (module X) where
 
-import Control.Monad
-import Control.Monad.Fix
-import Control.Monad.Ether.Implicit.Reader
-import Control.Monad.Ether.Implicit.Writer
-import Control.Monad.Ether.Implicit.State
-import Control.Monad.Ether.Implicit.Except
+import Control.Monad as X
+import Control.Monad.Fix as X
+import Control.Monad.Ether.Implicit.Reader as X
+import Control.Monad.Ether.Implicit.Writer as X
+import Control.Monad.Ether.Implicit.State as X
+import Control.Monad.Ether.Implicit.Except as X
+import Control.Monad.Trans.Ether.Dispatch as X hiding (pack, unpack)
+import Control.Ether.TH as X (ethereal)