+++ test/Regression/T2.hs
@@ -1,0 +1,8 @@
+module Regression.T2 (test2) where
+
+import Test.Tasty
+
+test2 :: TestTree
+test2 = testGroup "T2"
+  [
+  ]